[refactor] refactor code with new typing scheme.
