--- src/app/(BottomTabNavigation)/Cases/types.ts
@@ -1,9 +1,0 @@
-export type CaseId = number;
-export type UserUid = string;
-
-export type CaseCardProps = {
-  id: CaseId;
-  title: string;
-  status: string;
-  imageUrl?: string;
-};
